CI: 🧑‍🔧 auto WP update

+++ wordpress/scripts/provision.sh
@@ -72,7 +72,7 @@ if ! $WPCLI core is-installed --quiet &> /dev/null; then
 	echo "----------------------------------"
 	echo
 	if [ ! -z "${WORDPRESS_ENV}" ] && [ "${WORDPRESS_ENV}" = "dev" ]; then # we are on local dev environment (in docker)
-		echo $en "- Installing WordPress $ec"
+		echo $en "- Installing WordPress as localhost $ec"
 		$WPCLI core install --url="http://localhost" --title="superstack" --admin_user="superstack" --admin_password="superstack" --admin_email="tech+superstack@superhuit.ch" --quiet &> /dev/null
 		echo "✔"
 	elif [ ! -f "$WORDPRESS_PATH/p.txt" ]; then
@@ -92,23 +92,25 @@ if ! $WPCLI core is-installed --quiet &> /dev/null; then
 		[ -z "${WORDPRESS_URL}" ] && echo "ERROR: Please define WORDPRESS_URL environment variable" 1>&2 && exit 1
 		[ -z "${WORDPRESS_VERSION}" ] && echo "ERROR: Please define WORDPRESS_VERSION environment variable" 1>&2 && exit 1
 		# install
-		echo $en "- Installing WordPress $ec"
-		$WPCLI core download --version="$WORDPRESS_VERSION" --locale="$WORDPRESS_LOCALE"  --quiet &> /dev/null
-		$WPCLI config create --dbhost="$WORDPRESS_DB_HOST" --dbname="$WORDPRESS_DB_NAME" --dbuser="$WORDPRESS_DB_USER" --prompt=dbpass < $WORDPRESS_PATH/p.txt  --quiet &> /dev/null
-		$WPCLI core install --version="$WORDPRESS_VERSION" --locale="$WORDPRESS_LOCALE" --url="$WORDPRESS_URL" --title="$WORDPRESS_THEME_TITLE" --admin_user="$WORDPRESS_ADMIN_USER" --admin_email="$WORDPRESS_ADMIN_EMAIL"  --quiet &> /dev/null
-		rm $WORDPRESS_PATH/p.txt
-		echo "✔"
-	fi
-else
-	echo "- WordPress already installed"
-	# Check if current WordPress version matches requested version
-	CURRENT_VERSION=$($WPCLI core version --quiet)
-	if [ "$CURRENT_VERSION" != "$WORDPRESS_VERSION" ]; then
-		echo $en "- Updating WordPress from $CURRENT_VERSION to $WORDPRESS_VERSION $ec"
-		$WPCLI core update --version="$WORDPRESS_VERSION" --force --quiet &> /dev/null
-		echo "✔"
-	else
-		echo "- WordPress version $WORDPRESS_VERSION already installed"
+
+		if $WPCLI core is-installed --quiet; then
+			echo $en "- Installing WordPress $ec"
+			$WPCLI core download --version="$WORDPRESS_VERSION" --locale="$WORDPRESS_LOCALE"  --quiet &> /dev/null
+			$WPCLI config create --dbhost="$WORDPRESS_DB_HOST" --dbname="$WORDPRESS_DB_NAME" --dbuser="$WORDPRESS_DB_USER" --prompt=dbpass < $WORDPRESS_PATH/p.txt  --quiet &> /dev/null
+			$WPCLI core install --version="$WORDPRESS_VERSION" --locale="$WORDPRESS_LOCALE" --url="$WORDPRESS_URL" --title="$WORDPRESS_THEME_TITLE" --admin_user="$WORDPRESS_ADMIN_USER" --admin_email="$WORDPRESS_ADMIN_EMAIL"
+			rm $WORDPRESS_PATH/p.txt
+			echo "✔"
+		else
+			# Check if current WordPress version matches requested version
+			CURRENT_VERSION=$($WPCLI core version --quiet)
+			if [ "$CURRENT_VERSION" != "$WORDPRESS_VERSION" ]; then
+				echo $en "- Updating WordPress from $CURRENT_VERSION to $WORDPRESS_VERSION $ec"
+				$WPCLI core update --version="$WORDPRESS_VERSION" --force --quiet &> /dev/null
+				echo "✔"
+			else
+				echo "- WordPress version $WORDPRESS_VERSION already installed"
+			fi
+		fi
 	fi
 fi
